Moved torrent management endpoints to channels module
