--- Tribler/Core/Modules/restapi/channels_endpoint.py
@@ -1,64 +1,0 @@
-import json
-import time
-
-from twisted.web import http, resource
-
-from Tribler.Core.Modules.restapi import VOTE_SUBSCRIBE, VOTE_UNSUBSCRIBE
-from Tribler.Core.Modules.restapi.util import convert_db_channel_to_json, convert_db_torrent_to_json
-from Tribler.Core.simpledefs import NTFY_CHANNELCAST
-from Tribler.community.allchannel.community import AllChannelCommunity
-
-
-class ChannelTorrentsEndpoint(BaseChannelsEndpoint):
-    """
-    This endpoint is responsible for managing torrents in a channel.
-    """
-
-    def __init__(self, session, cid):
-        BaseChannelsEndpoint.__init__(self, session)
-        self.cid = cid
-
-    def render_GET(self, request):
-        """
-        .. http:get:: /channels/discovered/(string: channelid)/torrents
-
-        A GET request to this endpoint returns all discovered torrents in a specific channel. The size of the torrent is
-        in number of bytes. The last_tracker_check value will be 0 if we did not check the tracker state of the torrent
-        yet.
-
-            **Example request**:
-
-            .. sourcecode:: none
-
-                curl -X GET http://localhost:8085/channels/discovered/da69aaad39ccf468aba2ab9177d5f8d8160135e6/torrents
-
-            **Example response**:
-
-            .. sourcecode:: javascript
-
-                {
-                    "torrents": [{
-                        "id": 4,
-                        "infohash": "97d2d8f5d37e56cfaeaae151d55f05b077074779",
-                        "name": "Ubuntu-16.04-desktop-amd64",
-                        "size": 8592385,
-                        "category": "other",
-                        "num_seeders": 42,
-                        "num_leechers": 184,
-                        "last_tracker_check": 1463176959
-                    }, ...]
-                }
-
-            :statuscode 404: if the specified channel cannot be found.
-        """
-        channel_info = self.get_channel_from_db(self.cid)
-        if channel_info is None:
-            return ChannelTorrentsEndpoint.return_404(request)
-
-        torrent_db_columns = ['Torrent.torrent_id', 'infohash', 'Torrent.name', 'length',
-                              'Torrent.category', 'num_seeders', 'num_leechers', 'last_tracker_check']
-        results_local_torrents_channel = self.channel_db_handler\
-            .getTorrentsFromChannelId(channel_info[0], True, torrent_db_columns)
-
-        results_json = [convert_db_torrent_to_json(torrent_result) for torrent_result in results_local_torrents_channel]
-        return json.dumps({"torrents": results_json})
